UEFI http boot failure investigation.
Python http.server seems to FIN after the first HEAD request to size
the loaded file is completed and ACKed. What happens next is interesting.
On low latency connections, the GET request to download may get sent
after the server sends the FIN but before the client has a chance to
process it. The net result is:
- Server ignores GET
- HttpBootLoadFile returns EFI_CONNECTION_FIN. Boot fails.

In the other case, client handles the FIN before attempting the GET,
so there's a proper three-way handshake as part of GET.

Here you can find a capture showing both a good and bad http boots
(UEFI ran on a Socionext Synquacer, which is an edk2-platform).
HTTP server was started as "python3 -m http.server 1234".

Also reproed on RPi4 (also an edk2-platform). Presumably affects
any platform, but the problem is masked with latency (big LANs,
WAN, etc).

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>